added core and nexus config logic
